feat: add constant folder
