+++ container/src/shared/types/errors.ts
@@ -1,19 +1,4 @@
-export type ErrorType = 'network' | 'validation' | 'permission' | 'unknown' | 'component_crash';
-export type ErrorSeverity = 'critical' | 'warning' | 'info';
-
-export const ERROR_TYPE = {
-  NETWORK: 'network',
-  VALIDATION: 'validation',
-  PERMISSION: 'permission',
-  UNKNOWN: 'unknown',
-  COMPONENT_CRASH: 'component_crash',
-} as const;
-
-export const ERROR_SEVERITY = {
-  CRITICAL: 'critical',
-  WARNING: 'warning',
-  INFO: 'info',
-} as const;
+import { type ErrorSeverity, type ErrorType, ERROR_SEVERITY, ERROR_TYPE } from '@/shared/constants';
 
 export interface AppError extends Error {
   severity: ErrorSeverity;
